backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.19
pydantic==2.0.3
pydantic-extra-types==2.0.0
pydantic-core==2.6.3
email-validator==2.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
azure-storage-blob==12.17.0
pymysql==1.1.0
python-dotenv==1.0.0